Cleaning up  a bit
