+++ BlockChain/src/lt/viko/eif/blockChain/KeyOperations/SomethingLikeKeyChainAlternative.java
@@ -1,4 +1,4 @@
-package lt.viko.eif.blockChain.SingatureAndGui;
+package lt.viko.eif.blockChain.KeyOperations;
 
 import java.io.IOException;
 import java.nio.file.Files;
